Add GH Action for releases
This PR adds a GitHub Action to publish a new GitHub release everytime a change is merged to `main`. The release will also include the latest version of the tools in a tarball, which can easily be untarred to some place like `/usr/local/bin` to make system installs easy.